Move lib/sshutils/x11 to session/networking/x11
