+++ lib/service/servicecfg/ssh.go
@@ -21,8 +21,8 @@ package servicecfg
 import (
 	"github.com/gravitational/teleport/lib/limiter"
 	"github.com/gravitational/teleport/lib/services"
-	"github.com/gravitational/teleport/lib/sshutils/x11"
 	"github.com/gravitational/teleport/lib/utils"
+	"github.com/gravitational/teleport/session/networking/x11"
 )
 
 // SSHConfig configures Teleport's SSH service.
