+++ session/networking/x11/auth.go
